REFACTOR: Move user_model tests to app_user app
